Implement #126 and improve some settings
